Refactor import paths for ScheduleStartDate to use centralized components location
